Refactor Tado to use runtime_data
