+++ homeassistant/components/tado/services.py
@@ -12,7 +12,6 @@ from .const import (
     ATTR_MESSAGE,
     CONF_CONFIG_ENTRY,
     CONF_READING,
-    DATA,
     DOMAIN,
     SERVICE_ADD_METER_READING,
 )
@@ -40,7 +39,8 @@ def setup_services(hass: HomeAssistant) -> None:
         reading: int = call.data[CONF_READING]
         _LOGGER.debug("Add meter reading %s", reading)
 
-        tadoconnector = hass.data[DOMAIN][entry_id][DATA]
+        entry = hass.config_entries.async_get_entry(entry_id)
+        tadoconnector = entry.runtime_data.tadoconnector
         response: dict = await hass.async_add_executor_job(
             tadoconnector.set_meter_reading, call.data[CONF_READING]
         )
